+++ entry/src/main/ets/pages/socketAppenderConstants.ts
@@ -27,36 +27,35 @@ import {
 import { common } from '@kit.AbilityKit';
 
 export function InitializeAllLoggers(ctx: common.UIAbilityContext) {
-  // LogManager.interceptHilog();
-  // LogManager.setLogFilePath(ctx.filesDir);
-  // const socketAppender = new TCPSocketAppender({
-  //   address: '114.xxx.xxx.xxx',
-  //   port: 1234,
-  //   name: 'socket',
-  //   level: Level.ALL
-  // });
-  // const fileAppender_a = new FileAppender('log.txt', 'main', Level.ALL, {
-  //   useWorker: true,
-  //   maxFileSize: 1,
-  //   maxCacheCount: 2
-  // });
-  // const consoleAppender = new ConsoleAppender(Level.ALL)
-  //   .setLayout(new PatternLayout('%d%5L%5l%5p%r %C %% %m'))
-  // const fAppender = new FileAppender('Xlog.log', 'mainAppender', Level.ALL, {
-  //   useWorker: true
-  // }).setLayout(new PatternLayout('layout changed %m'));
-  // const dbAppender = new DatabaseAppender({
-  //   ctx,
-  //   name: 'mainDbAppender',
-  //   level: Level.ALL
-  // });
-  // LogManager
-  //   .registerLogger('Index')
-  //   .registerLogger('SplashPage')
-  //   .registerLoggers('SettingPage', 'AccountPage')
-  //   .bindAppenderGlobally(fAppender)
-  //   .bindAppenderGlobally(fileAppender_a)
-  //   .bindAppenderGlobally(consoleAppender)
-  //   .bindAppenderGlobally(socketAppender)
-  //   .bindAppenderGlobally(dbAppender)
+  LogManager.setLogFilePath(ctx.filesDir);
+  const socketAppender = new TCPSocketAppender({
+    address: '114.xxx.xxx.xxx',
+    port: 1234,
+    name: 'socket',
+    level: Level.ALL
+  });
+  const fileAppender_a = new FileAppender('log.txt', 'main', Level.ALL, {
+    useWorker: true,
+    maxFileSize: 1,
+    maxCacheCount: 2
+  });
+  const consoleAppender = new ConsoleAppender(Level.ALL)
+    .setLayout(new PatternLayout('%d%5L%5l%5p%r %C %% %m'))
+  const fAppender = new FileAppender('Xlog.log', 'mainAppender', Level.ALL, {
+    useWorker: true
+  }).setLayout(new PatternLayout('layout changed %m'));
+  const dbAppender = new DatabaseAppender({
+    ctx,
+    name: 'mainDbAppender',
+    level: Level.ALL
+  });
+  LogManager
+    .registerLogger('Index')
+    .registerLogger('SplashPage')
+    .registerLoggers('SettingPage', 'AccountPage')
+    .bindAppenderGlobally(fAppender)
+    .bindAppenderGlobally(fileAppender_a)
+    .bindAppenderGlobally(consoleAppender)
+    .bindAppenderGlobally(socketAppender)
+    .bindAppenderGlobally(dbAppender)
 }