+++ library/src/main/cpp/types/libliblog4a/index.d.ts
@@ -1,0 +1,7 @@
+import { Logger } from "../../../ets/Logger";
+
+export const getLogger: (ident: string) => Logger;
+
+export const hasLogger: (ident: string) => boolean;
+
+export const createLogger: (ident: string, logger: Logger) => void;